[WIP][TESTING] _set_symbolic_storage_offset on meta tensors
Previously, the symbolic storage offset is not actually used if the tensor is not a view. That's because the empty tensor construction (just above the change in meta_utils.py in this PR) doesn't take the storage offset as a parameter.

Note, this is not ready to land. There's some failures (e.g. in test_adv_index_batch) that need to be investigated. Submitting this to see CI.

ghstack-source-id: 3ee8ff775585c71751227ad08676a113576ab7c4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/114424