WIP: [inductor] Force IEEE-compliant division in triton
Fixes #101039

In the issue we have something like:
```python
(x / 1e-6) + (y / 1e-6)
```
which triton turns into
```python
fma(x, 1e6, y * 1e6)
```

Where we have one division calculated to float32 precision,
and the other calculated to infinite precision as part of the `fma`.
In the issue this leads to a bad normalization in softmax.

Here we explicity use the `div_rn` which forces IEEE compliant rounding.

ghstack-source-id: e48399334b3eb9f0ecbbefc02a658fae10a423ad
Pull Request resolved: https://github.com/pytorch/pytorch/pull/115435