Rename test_legacy_vmap.py TestCase names
The problem is that the dynamo_test_failures logic recognizes tests by
their TestClass.test_name. Unfortunately we have duplicate
TestClass.test_name in test_legacy_vmap and test_vmap. This PR
unduplicates them.

Something more robust would have been to include the test file name in
the dynamo_test_failures logic, but... it's a bit too late for that. We
can fix it if it becomes more of a problem in the future.

ghstack-source-id: 2b8d52e8b3d2821cae3676dcb564a519a16dfd94
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117320