Remove DICT_KEYS guard
This guard was trying to do too much. We replace it for the weaker
DICT_CONST_KEYS, which is only valid when all the values are constants
(i.e. easy to print out).

We also implement an optimisation where, if all the keys to a dictionary
are constant, we generate just one guard of the form
```
list(d.keys()) == ["key1", "key2"]
```

This is the type of guards we used to generate initially before this set
of PRs.

ghstack-source-id: fca1972f5ca07220ff5c06d4965001117d5dbc00
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118630