Reduce scope of dispatching in logcumsumexp_backward
Everything inside the `AT_DISPATCH` block is being compiled 5 times,
so it makes sense to limit it to the only line that uses `scalar_t` which is
the `numeric_limits` query.

Also a small optimization, instead of computing `grad.log()` and `(-grad).log()`
we can compute `grad.abs().log()` which is 2 pointwise ops instead of 3.

ghstack-source-id: cd2d9c5101e660185aa54808d288fba25484f7d0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119397