[aot_inductor] replace TORCH_CHECK with AOTI_TORCH_CHECK in the generate cpp code
In some cases where we have TORCH_CHECK in loops, it may cause the host
compiler to spend hours optimizing the run_impl function. This PR
mitigated the issue by replacing TORCH_CHECK with a custom AOTI_TORCH_CHECK,
where we place the underneath assert function in c_shim so that
it won't be inlined.

If forcing noinline caused any serious perf regression, we could
either add an option to turn on/off enable noinline. Or, we could
another an option to just turn AOTI_TORCH_CHECK into a no-op, similar
to the ```assert``` macro from cassert.

ghstack-source-id: 56452e0ab08523e7a6bff0e3ddd008d89fa06f1d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119220