[c10d] explicitly abort communicators in destroy_process_group call Summary: This PR tries to resolve issue #119215.
Basically,  processgroup shutdown (and hence ncclCommAbort) is called in
destroy_process_group APIs for the corresponding PGs. and in the
destructor of ProcessGroup, we avoid calling abort/ncclCommAbort.
Instead, it just checks if the users have explicitly already called destroy_process_group. If
not, Destructor will log a warning and encourage/expect users to do so
for cleanup of resources of PGs.

ghstack-source-id: 35bf3b66b852427d8799ad02f9c1b452c1ed390d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119250