[inductor] Refactor device guards Python codegen to allow nested indentation
Summary: The codegen of `with torch.cuda._DeviceGuard` context manager in the
Python wrapper code is implemented via `device_cm_stack: contextlib.ExitStack()`.
As the context managers in the stack are `code.indent()`, this means that the
whole stack is unindented at once on `device_cm_stack.close()`. This
becomes problematic when attempting to codegen indented code (e.g., for
control flow in Python and / or nested subgraph codegen-ing).

In this PR, we refactor the device guard codegen-ing in Python by
replacing the `device_cm_stack` by explicit indent and unindent calls
for entering and exiting the `with torch.cuda._DeviceGuard` context
manager. This allows for nested device guard context managers and better
aligns with other indented codegen-ing intertwined with it (e.g., for
nested subgraph codegen-ing).

This is necessary for the upcoming support for `torch.cond` (and other
control flow operators) in Inductor. Before that, the only change in
the Python wrapper codegen is that the `return outputs` is now happening
outside the `with torch.cuda._DeviceGuard` context manager.

Test Plan: CI

ghstack-source-id: b2916bea6ee90399dc5f65d466a63d77ed60d135
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119673