[inductor] allow padding mm/bmm/addmm in the presence of dynamic dims
Previously, pad_mm skips cases where any input tensor has symbolic
dimension or stride. This is too constraint in practise.
This PR enables this pass to pad non-symbolic dimensions in
the presence of dynamic dims. For example, with this PR, we could
pad the K dimension (i.e. 21) for torch.mm(A[s0, 21], B[20, 21]).

ghstack-source-id: 8d316f279c0c341a1f48c5c57106fbd9c3f74ccc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120073