[dtensor] simplify outputs wrapping handling
This PR simplifies the outputs wrapping handling in op dispatch, to make
it simpler and easier to understand.

It also enables a new case, where if the output DTensorSpec for the res is
None, and the res is a scalar tensor, we will just return the scalar
tensor instead of wrapping it with a DTensor.

ghstack-source-id: 1b740420549a79a95185e00a8c68b90466b50450
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120297