[Inductor Cutlass backend] Development feature flag
Introducing a feature flag for development purposes, which allows
to disable the CUTLASS backend entirely.

This flag is not intended to be used by users, it is for development
purposes to ensure a working codebase when features are incrementally
added over the course of multiple commits / PRs and it cannot
be ensured that the cutlass backend is functional in-between
incremental commits.

To disable the cutlass backend, set the private variable
cutlass_utils._DISABLE_CUTLASS_BACKEND to True

ghstack-source-id: 1debd7cad26b260440b95a390003edb42ebefa1f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119597