[Inductor cutlass backend] Experimental extended Cutlass op generator
So far we have relied on the "generator.py" from
third_party/cutlass/python/cutlass_library/ to generate Cutlass op
configurations to be tried in autotuning. These op configs are not
sufficient to ensure good performance and dtype coverage in all cases.

This diff introduces an extended variant of that generator.py,
provided both as a source file and as a diff ( so that it may be applied
to future versions of cutlass_library.generator )

Test Plan:
* CI
* python test/inductor/test_max_autotune.py

ghstack-source-id: c75c966f8a1c6b76cbd71c85cd808ee00fbc9e16
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119009