Replace TTIR string parsing with structured MLIR in Triton kernel  mutation analysis
Summary: Previously, we've relied on lark-based parsing of string TTIR
representation dumped by Triton compiler. However, this has proven to be
brittle in the face of changes both in the user-written Triton kernel
code and in the Triton compiler code.

In this PR, we add an alternative way of mining the function information
from the TTIR based on walking the tree of structured MLIR entities.
To this end, we rely on the MLIR bindings exposed by libtriton.

For now, we introduce gating based on whether
`ttir_module.hasattr("walk")`. This will allow switching to the newly
introduced approach to TTIR analysis only when the new MLIR bindings
become available in the Triton pin. Before then, we'll keep using the
old string TTIR parsing-based approach.

Test Plan: CI

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 35323e466c1f906cc3246be87129a8a96627cd5e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120476