Move attention kernels from meta_registrations to fake_impls
This PR is just code movement to make the code review easier. The final goal is to remove the xfails for some of the test_fake opinfos for these ops. The opinfos are failing because the outputs can have mixed devices - we need to move them to fake_impls first before we can support mixed device returns.

ghstack-source-id: 88a1027c32bbe84a424727bb8c20dcc7209b2c67
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120682