[AOTI] Store OpOverload in ir.ExternKernel
Summary: Currently the logics for filling the default value for optional arguments are scattered in several places. By storing OpOverload in the base ExternKernel class, we can simplify codegen_kwargs, and this is a preparation step for enabling the torchgen-ed C shim. codegen_args can also be simplified, but that can come later.

ghstack-source-id: 5310d00225c34ac43db99cd78f52d72f7ef0317a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120629